requirements.txt
anyio==3.4.0
asgiref==3.4.1
authy==2.2.6
backports.entry-points-selectable==1.1.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
defusedxml==0.7.1
distlib==0.3.4
dj-database-url==0.5.0
Django==4.0.2
django-allauth==0.49.0
django-cors-headers==3.11.0
django-environ==0.8.1
django-heroku==0.3.1
django-rest-auth==0.9.5
djangorestframework==3.13.1
# djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.1.0
dnspython==2.1.0
email-validator==1.1.3
fastapi==0.70.1
filelock==3.4.0
gunicorn==20.1.0
h11==0.12.0
httptools==0.2.0
idna==3.3
itsdangerous==2.0.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy==1.21.4
oauthlib==3.2.0
orjson==3.6.5
pipenv==2021.11.23
platformdirs==2.4.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycparser==2.21
pydantic==1.8.2
# PyJWT==2.4.0
python-decouple==3.6
python-dotenv==0.19.2
python-multipart==0.0.5
python3-openid==3.2.0
pytz==2021.3
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.1
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.2
starlette==0.16.0
twilio==7.9.1
typing_extensions==4.0.1
ujson==4.3.0
uritemplate==4.1.1
urllib3==1.26.7
uvicorn==0.15.0
uvloop==0.16.0
virtualenv==20.10.0
virtualenv-clone==0.5.7
watchgod==0.7
websockets==10.1
whitenoise==6.0.0
